pyproject.toml
[tool.poetry]
name = "media insights"
version = "1.0.0"
description = "Media Insights Project with Jessup Jong and Darren Colby"
authors = ["Jessup Jong and Darren Colby"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8.1"
google-api-python-client = "^2.76.0"
google-auth-oauthlib = "^0.8.0"
oauth2client = "^4.1.3"
altair = "^4.2.2"
pandas = "^1.5.3"
autocorrect = "^2.6.1"
translators = "^5.5.6"
psutil = "^5.9.4"
transformers = "^4.26.1"
vadersentiment = "^3.3.2"
spacy = "^3.5.0"
requests = "^2.28.2"
pillow = "^8.4.0"
wordcloud = "^1.8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
